EntryPoint/requirements.txt
pandas~=1.1.4
Flask~=1.1.2
Werkzeug~=1.0.1
passlib~=1.7.4
bcrypt~=3.2.0
gunicorn==20.0.4
google-cloud-storage
google-cloud-logging
google-cloud-tasks==2.2.0
